# backend/public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"Contact": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"message": {
						"$ref": "#/components/schemas/Message"
					}
				},
				"required": [
					"id",
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Template": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"path": {
						"type": "string"
					},
					"params": {
						"items": {
							"$ref": "#/components/schemas/TemplateParam"
						},
						"type": "array"
					},
					"message": {
						"$ref": "#/components/schemas/Message"
					}
				},
				"required": [
					"id",
					"name",
					"path",
					"params",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TemplateParam": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"templateId": {
						"type": "number",
						"format": "double"
					},
					"template": {
						"$ref": "#/components/schemas/Template"
					},
					"nameParam": {
						"type": "string"
					},
					"typeParam": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"templateId",
					"template",
					"nameParam",
					"typeParam"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Message": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"contactId": {
						"type": "number",
						"format": "double"
					},
					"contact": {
						"$ref": "#/components/schemas/Contact"
					},
					"templateId": {
						"type": "number",
						"format": "double"
					},
					"template": {
						"$ref": "#/components/schemas/Template"
					},
					"sendDate": {
						"type": "string"
					},
					"sendTime": {
						"type": "string"
					},
					"text": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"contactId",
					"contact",
					"templateId",
					"template",
					"sendDate",
					"sendTime",
					"text",
					"createdAt",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IContactPayload": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/contact": {
			"get": {
				"operationId": "GetContacts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Contact"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Контакты"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateContact",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Contact"
								}
							}
						}
					}
				},
				"tags": [
					"Контакты"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IContactPayload"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}